refactor: fix filename casing
